fix(compass-crud): scope the query builder to the tab, not the collection
Two tabs on the same collection shared one builder, because the rows and
the panel were kept in a map keyed by namespace. They are two different
queries and should not share anything.

Replaces that map with useTabState, which Compass already provides for
exactly this: state scoped to a workspace tab that survives the unmount
switching tabs causes. The documents list beside it already uses it for
scroll position and column widths.

This also removes a module level map of mutable state, and the tests
that covered it, since the behaviour is now the platform's rather than
something this component reimplements.
